refactor(messaging): remove legacy domain event dispatch infrastructure
- Eliminado DispatchDomainEventsInterceptor ya que la persistencia de eventos se gestiona ahora mediante PersistDomainEventsToOutbox.
- Eliminado EventualConsistencyMiddleware, ya no necesario gracias al patrón Outbox y procesamiento asincrónico mediante HostedService.
- Simplificación y mejora de la arquitectura, eliminando componentes obsoletos del flujo anterior.
