--- src/Infrastructure/Common/Middlewares/InfrastructureBuilderExtensions.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-
-namespace AgendaManager.Infrastructure.Common.Middlewares;
-
-public static class InfrastructureBuilderExtensions
-{
-    public static void UseInfrastructureMiddleware(this WebApplication app)
-    {
-        app.UseMiddleware<EventualConsistencyMiddleware>();
-    }
-}
