--- src/Infrastructure/Common/Persistence/Interceptors/DispatchDomainEventsInterceptor.cs
@@ -1,84 +1,0 @@
-﻿using AgendaManager.Domain.Common.Interfaces;
-using AgendaManager.Infrastructure.Common.Middlewares;
-using MediatR;
-using Microsoft.AspNetCore.Http;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Diagnostics;
-
-namespace AgendaManager.Infrastructure.Common.Persistence.Interceptors;
-
-public class DispatchDomainEventsInterceptor(IHttpContextAccessor httpContextAccessor, IPublisher publisher)
-    : SaveChangesInterceptor
-{
-    public override InterceptionResult<int> SavingChanges(DbContextEventData eventData, InterceptionResult<int> result)
-    {
-        DispatchDomainEvents(eventData.Context).GetAwaiter().GetResult();
-
-        return base.SavingChanges(eventData, result);
-    }
-
-    public override async ValueTask<InterceptionResult<int>> SavingChangesAsync(
-        DbContextEventData eventData,
-        InterceptionResult<int> result,
-        CancellationToken cancellationToken = default)
-    {
-        await DispatchDomainEvents(eventData.Context, cancellationToken);
-
-        return await base.SavingChangesAsync(eventData, result, cancellationToken);
-    }
-
-    private async Task DispatchDomainEvents(DbContext? context, CancellationToken cancellationToken = default)
-    {
-        if (context is null)
-        {
-            return;
-        }
-
-        var entities = context.ChangeTracker
-            .Entries<IEntity>()
-            .Where(e => e.Entity.DomainEvents.Count != 0)
-            .Select(e => e.Entity)
-            .ToArray();
-
-        var domainEvents = entities
-            .SelectMany(e => e.DomainEvents)
-            .ToList();
-
-        entities.ToList().ForEach(e => e.ClearDomainEvents());
-
-        if (IsUserWaitingOnline())
-        {
-            AddDomainEventsToOfflineProcessingQueue(domainEvents);
-        }
-        else
-        {
-            await PublishDomainEvents(domainEvents, cancellationToken);
-        }
-    }
-
-    private bool IsUserWaitingOnline()
-    {
-        return httpContextAccessor.HttpContext is not null;
-    }
-
-    private async Task PublishDomainEvents(List<IDomainEvent> domainEvents, CancellationToken cancellationToken)
-    {
-        foreach (var domainEvent in domainEvents)
-        {
-            await publisher.Publish(domainEvent, cancellationToken);
-        }
-    }
-
-    private void AddDomainEventsToOfflineProcessingQueue(List<IDomainEvent> domainEvents)
-    {
-        var domainEventsQueue = httpContextAccessor.HttpContext!.Items
-                .TryGetValue(EventualConsistencyMiddleware.DomainEventsQueueKey, out var value) &&
-            value is Queue<IDomainEvent> existingDomainEvents
-                ? existingDomainEvents
-                : new Queue<IDomainEvent>();
-
-        domainEvents.ForEach(domainEventsQueue.Enqueue);
-
-        httpContextAccessor.HttpContext!.Items[EventualConsistencyMiddleware.DomainEventsQueueKey] = domainEventsQueue;
-    }
-}
